requirements.txt
python-telegram-bot==21.5
openai==1.57.0
httpx==0.27.2
reportlab==4.1.0
Pillow==10.4.0
flask==3.0.3
psycopg2-binary==2.9.9
